requirements.txt
asgiref==3.8.1
attrs==24.2.0
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.10
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.2.0
dj-email-url==1.0.6
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
django-cache-url==3.4.5
django-cors-headers==4.4.0
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
drf-spectacular==0.27.2
environs==11.0.0
filelock==3.16.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
marshmallow==3.22.0
packaging==24.1
pip-tools==7.4.1
platformdirs==4.3.6
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==75.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.6
wheel==0.44.0
whitenoise==6.7.0
psycopg2-binary
pillow
django-debug-toolbar
django-taggit
markdown